Refactor the label and value detection for collections
The new detection mechanism is a bit less reliant on ActiveRecord
and provides some better defaults for common use cases like Array,
Integer, String oder Nil.

Additionally custom methods for label and value now get used for
finding selected values in a collection as well as for rendering
the collection.

Detection of methods based on work in Plataformatec SimpleForm.