Fix I18n errors in test environment/travis-ci
Depending on operating system, filesystem and environment the load
order of tests was changing leading to different I18n configurations
due to the load_path being set in the middle of the test suite.

The tests that uses the load path now cleans up after itself and resets
the load path to the default empty value.