Use `with_clean_env` in integration-template.rb
We want bundler to pickup the configuration of the new generated
application, not the current `BUNDLE_GEMFILE` and `BUNDLE_PATH`.
Previously, the `bundle_command` helper would use
`Bundler.with_clean_env` under the hood. However, it uses
`Bundler.with_original_env` since
https://github.com/rails/rails/commit/b1464e34a1aa2ef6e1c181972dca3a14eb65b291.

In this case, the original environment in TravisCI includes
`BUNDLE_GEMFILE=<repo_root>/gemfiles/rails_5.2.gemfile`, but we don't
want to use that Gemfile here, but the one in the generated application.
So, `Bundler.with_clean_env` is a better fit here since it completely
cleans up bundler-related environment variables and let them being
infered again from the current working directory. We _do_ want to reuse
the same bundler path for speed, so we save the previous value before
resetting the environment and pass that to `bundle install`.