Remove appraisals
Unfortunately, the default `appraisals` setup does not allow dependency
caching. Since it uses the same folder for all gemfiles, the cache and
bundler configuration for each gemfile gets messed up by the rest of the
gemfiles.

Appraisals does not really do much anyways, so it's easy to replace it
with a few custom ad-hoc tasks.