Increase minimum Rails version to 7.2 and Ruby 3.1
I was mistakenly confusing security patches with supported versions, which led to a lot of confusing rabit holes with Rails 7.1, which is confusingly both supports Ruby 2.7 in the gemfile, but has syntax only compatible with Ruby 3.1 in ActiveRecord.

So the supported base versions are Rails 7.2 and Ruby 3.1, in line with Rails itself.