Remove namespace from input_html_options
Now that we're properly passing namespace into form_for and fields_for, Rails' built-in form helpers like text_field will apply the namespace to inputs. However, we're already generating these through dom_id() for use here (and elsewhere), so we need to send a nil down intp the input helper so that the namespace isn't doubled-up.