Fix inferred maxlength for :integer columns
Postgres and MySQL column limits aren't directly usable as character lengths, as they're provided in byte lengths. When developers want to use a string input instead of a numeric input, this would result in a maxlen of 1 for an 8 bit column, rather than a maxlen of 3, 2 instead of 5 for a 16 bit, etc.

This translates the bits into characters.

Fixes #1377